+++ packages/shopify-app-session-storage-drizzle/src/drizzle.ts
@@ -1,2 +1,2 @@
-export {DrizzleSessionStoragePostgres} from './adapters/drizzle-pg.adapter';
+export {DrizzleSessionStoragePostgres} from './adapters/drizzle-postgres.adapter';
 export {DrizzleSessionStorageSQLite} from './adapters/drizzle-sqlite.adapter';
